infra: Delete the linux-blink-animation-use-time-delta bot
The bot is very red:
https://ci.chromium.org/ui/p/chromium/builders/ci/linux-blink-animation-use-time-delta

And has been failing in the same way until at least Oct of 2022:
https://ci.chromium.org/ui/p/chromium/builders/ci/linux-blink-animation-use-time-delta/30080/overview

So it appears to be providing very little value. So this deletes it to
free up resources + trim down configs. If it needs to be restored,
this CL can be reverted.

Bug: 40591169
Change-Id: I5541f9bf5722fc69836046015f862fdd7294d551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5393342
Reviewed-by: Weizhong Xia <weizhong@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1278001}
NOKEYCHECK=True
GitOrigin-RevId: 4fb1b5f2645e1c5fe1b2163db85c6fa3f64aeecf