Skylab builder: Clean up skylab variants of CrOS ASH builders
We had the multiple variants corresponding to the all boards in
`cros-skylab-variants.json`. This CL combines them into 2 common
variants: the public and private variants.

Previously the board, model, and dut_pool were specified in variants.
Now (with this CL), they are specified in `waterfall.pyl`.

This CL doesn't change the variants of the non-ash builders (lacros
builders and GPU builders.)

This CL is a clean-up and should not change any behavior.

      except for the variant names.

Bug: b/331640657
Test: manually check that `testing/buildbot/*.json` are not changed
Change-Id: I5c7ae17216afb773f4f8eeeb08c2aab3d012a381
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5210912
Commit-Queue: Yoshiki Iguchi <yoshiki@chromium.org>
Reviewed-by: Xinan Lin <linxinan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1279857}
NOKEYCHECK=True
GitOrigin-RevId: d1664ef5a19b16e16bb5fcf72ab49f53cce42c9e