[infra] Move the autoshard exceptions file.
In order for builders that set their targets in starlark to be
autosharded, the autoshard exceptions file needs to be moved to a
location where it can be consumed by the starlark. lucicfg doesn't allow
for reading files outside of the directory where the top-level script is
located, so the autoshard exceptions file needs to be located somewhere
in //infra/config. This just moves the file so that the starlark
configuration will eventually be able to support autosharding.

Bug: 332762804
Change-Id: Ief1bdb51bd73f5958a8096910de2c50d285b9cc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5422032
Reviewed-by: Stephanie Kim <kimstephanie@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1282216}
NOKEYCHECK=True
GitOrigin-RevId: 4999e9796e68f5687b15a9ff9a6501da7eecba83