Add an experiment to run DIPS queries on a foreground sequence
The DIPS database queries are currently executed on a background
sequence.

This CL is adding a feature to allow to make an Arm in the SqlWalMode
study. The goal is to study the impact of background sequence and
WAL mode.

This feature is intended to study performance and won't be shipped to
stable. The study is only targeting Windows since it's the only
platform that do support low IO priority.

Bug: 40256732
Change-Id: Ib65922b481378ceb4fe4bac985da6a7f9d42477d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5392123
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Joshua Hood <jdh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284643}
NOKEYCHECK=True
GitOrigin-RevId: 27f30e14a2cebf03dcaf4e2efb3c582f87657c73