Optimize XRW origin trial checking flow to avoid UI thread hop
I was looking through some traces of WebView navigation, and saw there
could be quite a long delay starting the request while waiting for the
CheckXrwOriginTrialAsync() call to finish. This call has to post a task
to the UI thread, then post the result back to the IO thread before the
request can start. Posting tasks to the UI thread during navigation can
cause large delays if the UI thread is busy (especially on Android where
the system may be using the main thread).

This change attempts to eliminate the UI thread hop in most cases by
getting the XRW origin trial status as part of a navigation throttle.
Unfortunately, I didn't find a great way to pass this info to
AwProxyingURLLoaderFactory, so it's stored in a map keyed by the unique
navigation ID, which can then be grabbed when needed.

Bug: 332706093
Change-Id: I8bca92d2d3d26c06f9e5ad7f7698dea8683b5daf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5427274
Commit-Queue: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1284650}
NOKEYCHECK=True
GitOrigin-RevId: a2b60d4f89a32d711adcda1d029d6a6286d78391