Extend the timeout for leveldb_unittests on dbg builds
The Issue320.Test test in leveldb_unittests is time-sensitive and may
not finish within the 45s timeout on some dbg builds.

Instead of adjusting the kTimeoutMultiplier for all dbg builds (as we do
not observe this issue for other test suites and leveldb_unittests is a
third-party unit test suite), this CL specifically increases the timeout
for leveldb_unittests.

Bug: 40282321
Change-Id: I29eaaade7a1913a82421f21a24cb4e725c7038c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5441111
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Commit-Queue: Kuan Huang <kuanhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1285376}
NOKEYCHECK=True
GitOrigin-RevId: 0dd56cb567376edfe39019a5ccff4c3eaae2fc95