[infra] Enable compound suites to be used as bundles.
This allows a legacy compound suite to be included in a bundle or
referenced directly by a builder. Along with that, support was added for
removing a test from a bundle by using targets.remove as the per-test
modification for a test.

Bug: 40258588, 40666862
Change-Id: I24542595fdca8a7c9275e9e138781d9540ef01ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5415256
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1288319}
NOKEYCHECK=True
GitOrigin-RevId: 9766d3bf3a7744b93ea53a040141cbf1ea91e342