Disable AutofillParsingPatternProvider on ToT
We were planning to launch this experiment. But since launching from 1%
to 100% directly is not allowed on WebView, this CL disables the feature
again, before ramping the rollout up to 10% stable.

The CL also re-adds the field trial config entry that was removed when
enabling the flag by default.

Bug: b/40280853
Change-Id: I944aedf554404b6cc58bb1f69af7e1b749448687
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5466295
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Norge Vizcay <vizcay@google.com>
Cr-Commit-Position: refs/heads/main@{#1290547}
NOKEYCHECK=True
GitOrigin-RevId: 9c21ecdd15664cb787ad4b95b21d6fbbc403ccba