[infra] Enable android-specific behavior for tests in starlark.
This change adds support for builders specifying an os type of Android.
This can be specified via the targets_setting argument, which takes an
instance of targets.settings. The settings enable setting values that
control the expansion of all tests for a builder. The supported
attributes right now are:

* os_type - The type of OS that the tests are targeting, which controls
  the expansion of OS-specific args and swarming values (not yet
  implemented). The only supported value right now is
  targets.os_type.ANDROID, which will add details that are common to all
  Android tests.
* results_bucket - The bucket to upload results to for the
  Android-specific test presentation.
* use_swarming - Whether tests on the builder should be swarmed.

The builders migrated as part of this change contain the only uses of
the oreo_mr1_fleet. To avoid generate_buildbot_json.py from raising an
error due to the mixin not being referenced, targets.mixin was updated
with a generate_pyl_entry argument. By default all named mixins will
have an entry generated in mixins.pyl but if generate_pyl_entry is set
to False then an entry won't be generated for the mixin.

Bug: 40258588
Change-Id: Ida6103e5a69c14207247595fe19ed2d15bcddd52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5464280
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1290874}
NOKEYCHECK=True
GitOrigin-RevId: 1df0acc09e58445c306fd35da37a7ebd452628d6