[Adaptive Toolbar] Enable Translate and Bookmark buttons by default
This CL changes the AdaptiveButtonInTopToolbarTranslate, AdaptiveButtonInTopToolbarAddToBookmarks and SegmentationPlatformAdaptiveToolbarV2Feature flags to be enabled by default.

The translate and bookmarks flags will expose the two new buttons in settings, and SegmentationPlatformAdaptiveToolbarV2Feature will enable the multi-output segmentation model that can pick those 2 new buttons.

The multi-output code path had an issue that showed the buttons
even if the model was not available. This CL fixes this issue and
modifies a param to make this behavior intentional. This means that
the adaptive toolbar will appear on the first run, without waiting
for data collection.

Bug: b/283326658
Change-Id: I76d1c7393c83acb92785ffc6763918d81fe805b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463020
Reviewed-by: Siddhartha S <ssid@chromium.org>
Reviewed-by: Ritika Gupta <ritikagup@google.com>
Reviewed-by: Shakti Sahu <shaktisahu@chromium.org>
Commit-Queue: Salvador Guerrero Ramos <salg@google.com>
Cr-Commit-Position: refs/heads/main@{#1291621}
NOKEYCHECK=True
GitOrigin-RevId: f1f620317583ed037aa11b2c71edf75adb534f22