Add GPU Android workaround, refactor args
Adds the new --initial-find-device-attempts arg to all non-WebGPU
Telemetry-based GPU integration tests in order to work around flakes
caused by devices briefly going offline, particularly on certain devices
like Samsung S23s.

Since we have cases of other args we want applied everywhere in addition
to this, this also refactors the test definitions in basic_suites.star
use a new mixin, which auto-adds arguments that we want applied to
all GPU integration tests. This does result in the addition of a bunch
of --jobs arguments to suites that didn't have it before, but these are
a no-op since the underlying suite must have parallelization enabled for
the argument to have an effect.

Bug: 333414298
Change-Id: I405e642e17949f13f096e466dab5c7cb107e6e19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5474739
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292018}
NOKEYCHECK=True
GitOrigin-RevId: 774cd7c603e550c1e12f4f46aa76290c798f0b3d