Retire CompositeBGColorAnimationExp4 field trial
The field trial config change for OptimizeStaticAnimationProperties
flagged metrics regressions for background color animations. The
background color animations are conditionally run on the compositor in
experimental and are not ready for launch, as further optimization work
is required.

Previously, any animation with a property that is not supported on the
compositor would force the animation to run on main. With the static
property optimization, these animations can now run on the compositor.
If a background color animation is triggered via the background
shorthand, the animation will include static properties. Native paint
worklets do not fully integrate this optimization.

We plan to relaunch compositing background color via Finch, and will
create a new experiment once the optimization work is complete and
are ready to launch.

Bug: 336596362
Change-Id: I06006c0e964b27822d6887a601091c5928e56b17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5491143
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1295474}
NOKEYCHECK=True
GitOrigin-RevId: 3e77684804d14b6fbebe112d7ed53d6327da12d9