[infra] Sync test-only owners from //testing/buildbot to //infra/config.
The generated .json files in //testing/buildbot have additional owners
besides the general infra owners which enables CLs that only modify
tests for some builders to be approved without requiring a general infra
owner. This duplicates the ownership to the generated json files in
//infra/config/generated. This will allow test owners to approve a CL if the only starlark files modified are in //infra/configs/targets. Unfortunately, modifications to a builder's tests can't be approved by
without an infra owner because the starlark files for a builder group
are not (and should not be) owned by everyone. There's still some
benefit in that case because they will still appear as owners in the
gerrit owners box, so infra owners could easily request a review from a
test owner if necessary.

Bug: 40258588
Change-Id: I166ea634712ab7d93412ca20167c487b1508b331
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5512689
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296173}
NOKEYCHECK=True
GitOrigin-RevId: d136bbac74f46e0f5522730889388985f97e6aef