Enable WinSboxNoFakeGdiInit testing config
The feature is passing CQ and is ready for experimenting in canary.

With WinSboxNoFakeGdiInit enabled, gdi32.dll and other dlls which
depend on it (like user32.dll, shell32.dll...) will fail to load in
sandboxed processes with win32k lockdown enabled, as we are getting rid
of the shims on these DLLs which allow them to load anyway.

Bug: 326277735
Change-Id: Ia503472412dd611f7e4df8cb6e0e87930c61273d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5463616
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Auto-Submit: Stefan Smolen <ssmole@microsoft.com>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296333}
NOKEYCHECK=True
GitOrigin-RevId: 5ea1f7034a8279c2e58226498a351c6220501149