Add HistoryEmbeddings to fieldtrial_testing_config.json
Needed to support any Variations configuration.

This CL is a partial reland of this revert:
https://chromium-review.googlesource.com/c/chromium/src/+/5514223

As for the revert, I couldn't reproduce the test failure locally, but
made some modifications to the test to try to avoid this failure mode
on the buildbot.

Bug: b/338245900
Change-Id: Icb06212727cba90653b50b33e460da7b27f0601e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5515642
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296923}
NOKEYCHECK=True
GitOrigin-RevId: b7a33f8d933bb8209aeda186f79fba55ec06b4ee