Introduce the performance notification button
This CL introduces the performance notification toolbar button.
Currently, the button is permanently shown on the toolbar button and
doesn't do anything when it is clicked. Logic to trigger the toolbar
button to show/hide and show a dialog when clicked will be added in
follow up CLs.

Picture of Button:
https://screenshot.googleplex.com/8mKkodVgQhXe3vv.png

Bug: 338072465
Change-Id: I118afd10a84a31b4959d9a42042d4d46a650b38a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5503223
Reviewed-by: David Pennington <dpenning@chromium.org>
Commit-Queue: Steven Luong <stluong@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1296954}
NOKEYCHECK=True
GitOrigin-RevId: c723495362e67d875ff2a67fe7a77353c3c0e6ac