[DoH] Remove kDnsOverHttps feature flag
Removes kDnsOverHttps and its feature params. I couldn't think of a
compelling reason for why we would still need this since the
functionality is launched for all platforms (since iOS doesn't use
code in chrome/ AFAICT, and since we don't support Chrome for
Fuchsia anymore). Also, we still have the kDnsOverHttpsUpgrade flag
if we want to disable that functionality, and we can also disable
the DoH auto-upgrade behavior on a per-provider basis. I don't think
keeping this feature flag around as a killswitch makes sense either,
since using it would have security/privacy ramifications for users
that specifically opted in to the protections of DoH.

Bug: b:338033843
Change-Id: I778636345eaf8a7f7abed94482c579bc24f0fef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5526034
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Rainhard Findling <rainhard@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1300739}
NOKEYCHECK=True
GitOrigin-RevId: c40b0bab6871a54b741268bc0911f136c4a3ecd5