[Protected Audience] Support multi-threaded SellerWorklet
Why: This aims to reduce auction running time. We'll conduct a
Finch experiment to measure the performance impact.

How:
- Update `g_seller_instance` static variable to `g_seller_instances,
which is a vector of `V8HelperHolder`s.
- Use a round-robin scheduling and the state is shared among
worklet services in the same process (i.e. the same scope for
the threading).
- Configure fieldtrial_testing_config.json. This will also let
the browsertests and web_tests test the multi-threaded configuration.

Bug: 335613091
Change-Id: Ief03876e23ff51212bbe37a840bb422f459427dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5538750
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1302755}
NOKEYCHECK=True
GitOrigin-RevId: 85213af8632caaa69d9bbe9f6f42441956a04c4d