Enable SafeBrowsingNewGmsApi flags by default on Clank.
The regression on Android startup metric is addressed with the dummy
call on startup, so we are ready to proceed to 100%. See
http://b/314272272#comment30 for details.

This CL enables the following flags on Clank:
SafeBrowsingCallNewGmsApiOnStartup
SafeBrowsingNewGmsApiForBrowseUrlDatabaseCheck
SafeBrowsingNewGmsApiForSubresourceFilterCheck

Overriding SafeBrowsingNewGmsApiForBrowseUrlDatabaseCheck to disable
on WebView, because it is still being rolled out. The other two flags
don't apply to WebView.

Bug: 40935425,40938388
Change-Id: I3cc72e5c5d86a8f13d0ccfa229f36e5fe6331bad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5543081
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1303361}
NOKEYCHECK=True
GitOrigin-RevId: b6b9c9f56d05cd8b05ea74f2121efe9653457360