Enable migrate stale tabs in TabState FlatBuffer field trial
For safety the experiment was broken up into the core
of the migration and migrating stale Tabs (Tabs which go
uninteracted with, don't hit the usual save path (which
migrates the Tabs). There have been no issues identified
with the core of the migration so enabling migrate stale
Tabs as well.

Bug: b:307736600
Change-Id: I31aa94751d5ad4accb6899b6f7f18e58037478ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5570932
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1305855}
NOKEYCHECK=True
GitOrigin-RevId: 1b35c32849c3e5b7e68fef12ce46bb1cf71a6324