[DIPS] Add a new base::Feature to toggle DIPS database prepopulation.
Tests involving DIPS currently need to call
DIPSService::WaitForInitCompleteForTesting() to prevent flakiness.

We don't want to expose that in the public API when we move to
//content.

Instead, add an enabled-by-default Feature for prepopulation, and disable it in tests that don't want it.

We will use the same Feature in a future Finch experiment to confirm
it's safe to remove prepopulation entirely. Disable it in
fieldtrial_testing_config.json now to get ready.

Bug: 336563548
Change-Id: I13a80e2de35b28366a98b5ff3800ed9be792d1a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5559129
Reviewed-by: Joshua Hood <jdh@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Ryan Tarpine <rtarpine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1305957}
NOKEYCHECK=True
GitOrigin-RevId: 00b8da38e09000fb0461b9d47363a33007253606