Re-introduce fieldtrial for composited background color animations
In the process of perf-testing composited background color, we found
regressions in some of the tough cases.  It turns out that save layers
is expensive as it is likely not optimized out in Skia. Redundant save
layers are likely optimized out for normal rendered content, but paint
worklets are handled separately. It looks like the save layer is not
required for our native paint worklets.  Making saveLayers conditional
on run a JavaScript paint worklet.

Change-Id: Id74b2d9dc2e6e26e3090433feedd3be59f9b556b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5575974
Reviewed-by: Victor Miura <vmiura@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1308018}
NOKEYCHECK=True
GitOrigin-RevId: 517a725e23cd724c012d68b78de078a7c68abc3a