fuzztest: fix corpus_data requirement
This new fuzztest feature allows for saving corpuses to a database which
defaults to ~/.cache/fuzztest/. This feature requires having a limited
fuzzing time so that it can save corpuses at the end. As for now, we're
not using this feature so we can just leave this field empty.

Bug: 343688205
Change-Id: Id592cd659c1cd25d5ea91900f4ada155319b092f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5587876
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Paul Semel <paulsemel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1308732}
NOKEYCHECK=True
GitOrigin-RevId: ddefa5db38896b2b49b7b72e3994ba8c456624f7