Register the App-Bound key provider by default on Windows.
This CL enables the App-Bound key provider registration by default
on Windows.

This is needed as a pre-requisite for enabling encryption, as the
provider should remain registered if any data have been encrypted
with the key provided.

Do not revert this CL without consultation as features that
depend on this CL might also need to be reverted and/or turned
off to avoid data loss.

BUG=40227925

Change-Id: Ie583fc54a4934187eac8de731f9f07a8329d8d70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5551182
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1309453}
NOKEYCHECK=True
GitOrigin-RevId: 2d348854a394c05d7086a562479ed4ddb8e55a17