[v8] Prepare parameters for next iteration on finch experiment
The initial finch experiment is finished and the parameters were chosen
according to the data.
The next iteration of the finch experiment is supposed to verify the
local maximum in the three-dimensional parameter space by modifying each
individual parameter and checking the effect on the different metrics.

This CL sets the defaults of the feature parameters to the current
defaults in V8 (see https://crrev.com/c/5412938) and prepares the field
trial config for the next iteration of the experiment.

R=leszeks@chromium.org

Bug: 40945417
Change-Id: I82b637ec59c0a7a4e426224bcd1f96293f5e67e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5600266
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1310646}
NOKEYCHECK=True
GitOrigin-RevId: 859c9697d83f21b336ba6c5845b2fadde8f71ca5